Filter data for items that shouldn't have it and filter wool.
We used to fall Item.filterData() for this but that method is meant for
converting item data to block data during placement and does the wrong
thing for this case. Instead we just see if the item should have data and
if not set it to zero. We also have to filter wool data explicitly because
clients crash when given invalid wool data.

Remapped by Srg2Source from Bukkit/CraftBukkit@80d8c9a20ece0e4fb5948f850ff2067dc3f78258